I have finally tidied up the code and fixed a couple of bugs.
I have now implemented a filter spinner which the usr can add filters onto the image and tweak it. I have also added icons to make the options much clearer.

Somethings i would of liked to have done if i had the chance would be to add
a sharpen, brightness and saturation filters.